pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mdify"
version = "0.1.0"
description = "PDF → Markdown converter powered by local Ollama vision models"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Julius Pinsker" },
]
keywords = ["pdf", "markdown", "ocr", "vlm", "ollama", "qwen"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Scientific/Engineering",
    "Typing :: Typed",
]
dependencies = [
    "PyMuPDF>=1.24.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
]

[project.scripts]
mdify = "mdify.cli:main"

[project.urls]
Homepage = "https://github.com/jupinsker/mdify"
Repository = "https://github.com/jupinsker/mdify"
Issues = "https://github.com/jupinsker/mdify/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/mdify"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
